tools: gen openapi schema & zod schema
We don't need the full OpenAPI spec, there are a bunch of endpoints that
we just won't need. We can incrementally add them as we add features.
Wecan output the schema as a json file that we can use for the
`/openapi` endpoint.

Additionally, we can generate zod validation schemas based on the
openapi schema.