store: create a pouchdb store for the application
Allow the user to provide a path to the composes store, if none
is provided we fall back to either an environment variable or
`/var/lib/osbuild-composer/artifacts'. The pouchdb store will be
created inside the provided store directory.

This commit is in preparation for following commits where we will
use the store to manage composes and blueprints.