queue: add a worker queue
Add an initial queue class that delegates the work of building a compose
to a worker thread. For now the worker only creates the directory and
saves the request to the store. Calling the `image-builder-cli` will
come in later commits, since we need to add conversion of `hosted`
blueprint customizations to `on-prem` blueprint first.