main: extract app creation and add some factory methods
The service was being initialized on every request, this was fine until
we started adding event listeners. This caused a memory leak with a new
event listener being registered on every request. Instead, let's create
the store, queue & composes service on app creation with the help of
some factory methods which helps us bootstrap the application. This
change also helps with dependency injection which is particularly useful
for testing.