queue: refactor the queue to build manifests
We can add a manifest builder worker. While we may not need this for the
service, it is will be very useful for integration tests in the app,
since we can build manifests instead of full images and compare the
results to a known good manifest. We also add a factory method here for
creating the type of worker we need.