api/distributions: add the image type endpoint
Add and endpoint and tests for getting the list of image types available
for a specified distro. The matrix of supported image types for each
distro in CRC is quite complex. This commit simply takes the output from
image-builder-cli and filters this against the image types that
image-builder-crc generally supports. In practice, this means that the
shim will be able to support more image-types than what is currently
available in image-builder-crc.