multi: stricter typescript types
Adjust the typescript config to have stricter types and adjust the
linter to account for this.

The primary reason for the change was so that the frontend and the shim
have matching types. Since `exactOptionalPropertyTypes` is set in the
frontend.

Additionally, if we are importing a type, we should be explicit and
include the `type` keyword in the import statement.