tools: gen openapi schema & zod schema
We don't need the full OpenAPI spec for the shim, since there are some
endpoints that only make sense for the hosted service. We can filter out
the endpoints and types and if we need more endpoints, we can add them
at a later stage.

We will also need to pull down the cloudapi OpenAPI spec
from the osbuild-composer project. The same principle applies here, we don't
need the full spec, so we can filter this too.

Additionally, we can generate zod validation schemas based on the
openapi schemas for both sets of specs.