tools: gen openapi schema & zod schema
We don't need the full OpenAPI spec for the shim, since there are some
endpoints that only make sense for the hosted service. We can filter out
the endpoints and types and if we need more endpoints, we can add them
at a later stage.

Additionally, we can generate zod validation schemas based on the
openapi schema.