api/composes: add GET /composes endpoint
As a first step to add compose functionality to the application. We will
first add the GET composes endpoint and we will add the necessary helper
functions and classes to do this. We are using models, services and
route handlers to be able to do this. We have also added a paginate
helper function, since this will be re-usable for other endpoints in
the application.